+++ spring-boot/jdbc/src/test/kotlin/io/bluetape4k/spring/data/exposed/jdbc/support/JdbcFluentQueryPlanTest.kt
@@ -4,7 +4,7 @@ import io.bluetape4k.assertions.assertFailsWith
 import io.bluetape4k.assertions.shouldBeEqualTo
 import io.bluetape4k.assertions.shouldBeFalse
 import io.bluetape4k.assertions.shouldBeTrue
-import io.bluetape4k.spring.data.exposed.jdbc.mapping.ExposedMappingContext
+import io.bluetape4k.spring.data.exposed.common.mapping.ExposedMappingContext
 import io.bluetape4k.spring.data.exposed.jdbc.repository.support.JdbcFluentQueryPlan
 import io.bluetape4k.spring.data.exposed.jdbc.repository.support.JdbcFluentQueryScope
 import org.junit.jupiter.api.Test