+++ spring-boot/r2dbc/src/main/kotlin/io/bluetape4k/spring/data/exposed/r2dbc/repository/query/PartTreeExposedR2dbcQuery.kt
@@ -1,9 +1,9 @@
 package io.bluetape4k.spring.data.exposed.r2dbc.repository.query
 
 import io.bluetape4k.logging.coroutines.KLoggingChannel
-import io.bluetape4k.spring.data.exposed.jdbc.repository.query.ExposedQueryCreator
-import io.bluetape4k.spring.data.exposed.jdbc.repository.query.ParameterMetadataProvider
-import io.bluetape4k.spring.data.exposed.jdbc.repository.support.toExposedOrderBy
+import io.bluetape4k.spring.data.exposed.common.repository.query.ExposedQueryCreator
+import io.bluetape4k.spring.data.exposed.common.repository.query.ParameterMetadataProvider
+import io.bluetape4k.spring.data.exposed.common.repository.support.toExposedOrderBy
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.channelFlow
 import org.jetbrains.exposed.v1.core.Op